Bash: Add env var for dotfiles-openjck repo
This environment variable is for the dotfiles-openjck repo that is
checked out in the ~/devel/repos directory, not the version that is
managed by vcsh. The former allows easy editing of README.md and
LICENSE, but the latter does not because they are hidden from the home
directory via sparse-checkout.