Bash: Don't fall back to empty string in config
I played around with setting -u across all Bash configuration files,
which would make this a requirement, and which would also require this
technique elsewhere. However, it broke Bash completion, which I can't
alter. I've come to feel it doesn't make sense to use "strict mode" or
anything like it in config files. Perhaps it can be enabled temporarily
in some functions, but not elsewhere.