Zellij: Load Zellij, not tmux, by default
I like Zellij because it provides somewhat better user experience than
tmux (e.g., common commands are documented on screen) and because it has
some features (e.g., floating panes) that tmux does not have.

For now, this is just an experiment. I might switch back to tmux.

The zellij function is removed for two reasons. First, it shadows the
built-in zellij command, which can be problematic at times. Second, it's
not really needed now that Zellij is starting up automatically. I'd
still like the default session to be named "main", but I haven't worked
that out yet.