Bash: Revert to using "nvim" for "n"
I've decided not to switch to Helix at this time because I'd wait for it
to mature longer. For example, at the moment, it does not support code
folding or EditorConfig, and the vscode extension for Helix is lacking.
I still think Helix is great, though, and I'd like to use it more in
the future.