Neovim: Separate maps for project / cwd files
The benefit with this approach is that <Leader>fp can now be used to
find all files in the current repo regardless of what the current
working directory is (not how it worked before), whereas <Leader>fc
shows only files in the current working directory. The latter can be
useful in directories that are not part of a repo.